Refactor: Rename BaseEntity
